+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `counter_obj:collect()`;
   - `gauge_obj:collect()`;
   - `histogram_obj:collect()`;
+  - `summary_obj:collect()`;
 
 ### Changed
 - Setup cartridge hotreload inside the role